FASTAPI/app/requirements.txt
fastapi 
uvicorn 
requests
python-dotenv
python-multipart
psycopg2
neo4j
sqlalchemy
pandas
openpyxl
langchain-community